refactor: Footer, Form  js -> ts
